Refactor with MeasurementAgent.
