+++ src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha/ReportsService.kt
@@ -15,14 +15,10 @@
 package org.wfanet.measurement.reporting.service.api.v1alpha
 
 import com.google.protobuf.ByteString
-import com.google.protobuf.Duration
-import com.google.protobuf.duration
-import com.google.protobuf.util.Durations
 import com.google.protobuf.util.Timestamps
 import io.grpc.Status
 import io.grpc.StatusException
 import java.io.File
-import java.security.PrivateKey
 import java.security.SecureRandom
 import java.time.Instant
 import kotlin.math.min
@@ -34,46 +30,16 @@ import org.wfanet.measurement.api.v2.alpha.ListReportsPageTokenKt.previousPageEn
 import org.wfanet.measurement.api.v2.alpha.copy
 import org.wfanet.measurement.api.v2.alpha.listReportsPageToken
 import org.wfanet.measurement.api.v2alpha.CertificatesGrpcKt.CertificatesCoroutineStub
-import org.wfanet.measurement.api.v2alpha.CreateMeasurementRequest
-import org.wfanet.measurement.api.v2alpha.DataProviderKey
 import org.wfanet.measurement.api.v2alpha.DataProvidersGrpcKt.DataProvidersCoroutineStub
-import org.wfanet.measurement.api.v2alpha.EncryptionPublicKey
 import org.wfanet.measurement.api.v2alpha.Measurement
-import org.wfanet.measurement.api.v2alpha.Measurement.DataProviderEntry
-import org.wfanet.measurement.api.v2alpha.MeasurementConsumer
 import org.wfanet.measurement.api.v2alpha.MeasurementConsumerKey
 import org.wfanet.measurement.api.v2alpha.MeasurementConsumersGrpcKt.MeasurementConsumersCoroutineStub
-import org.wfanet.measurement.api.v2alpha.MeasurementKey
-import org.wfanet.measurement.api.v2alpha.MeasurementKt.DataProviderEntryKt.value as dataProviderEntryValue
-import org.wfanet.measurement.api.v2alpha.MeasurementKt.dataProviderEntry
-import org.wfanet.measurement.api.v2alpha.MeasurementSpec
-import org.wfanet.measurement.api.v2alpha.MeasurementSpec.VidSamplingInterval
-import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.duration as measurementSpecDuration
-import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.impression as measurementSpecImpression
-import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.reachAndFrequency as measurementSpecReachAndFrequency
-import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.vidSamplingInterval
 import org.wfanet.measurement.api.v2alpha.MeasurementsGrpcKt.MeasurementsCoroutineStub
-import org.wfanet.measurement.api.v2alpha.RequisitionSpec.EventGroupEntry
-import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt.EventGroupEntryKt.value as eventGroupEntryValue
-import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt.eventFilter as requisitionSpecEventFilter
-import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt.eventGroupEntry
-import org.wfanet.measurement.api.v2alpha.TimeInterval as MeasurementTimeInterval
-import org.wfanet.measurement.api.v2alpha.createMeasurementRequest
-import org.wfanet.measurement.api.v2alpha.differentialPrivacyParams
 import org.wfanet.measurement.api.v2alpha.getCertificateRequest
-import org.wfanet.measurement.api.v2alpha.getDataProviderRequest
 import org.wfanet.measurement.api.v2alpha.getMeasurementConsumerRequest
-import org.wfanet.measurement.api.v2alpha.getMeasurementRequest
-import org.wfanet.measurement.api.v2alpha.measurement
-import org.wfanet.measurement.api.v2alpha.measurementSpec
-import org.wfanet.measurement.api.v2alpha.requisitionSpec
-import org.wfanet.measurement.api.v2alpha.timeInterval as measurementTimeInterval
 import org.wfanet.measurement.api.withAuthenticationKey
 import org.wfanet.measurement.common.base64UrlDecode
 import org.wfanet.measurement.common.base64UrlEncode
-import org.wfanet.measurement.common.crypto.PrivateKeyHandle
-import org.wfanet.measurement.common.crypto.SigningKeyHandle
-import org.wfanet.measurement.common.crypto.hashSha256
 import org.wfanet.measurement.common.crypto.readCertificate
 import org.wfanet.measurement.common.crypto.readPrivateKey
 import org.wfanet.measurement.common.grpc.failGrpc
@@ -82,22 +48,10 @@ import org.wfanet.measurement.common.grpc.grpcRequireNotNull
 import org.wfanet.measurement.common.identity.apiIdToExternalId
 import org.wfanet.measurement.common.identity.externalIdToApiId
 import org.wfanet.measurement.common.readByteString
-import org.wfanet.measurement.consent.client.measurementconsumer.decryptResult
-import org.wfanet.measurement.consent.client.measurementconsumer.encryptRequisitionSpec
-import org.wfanet.measurement.consent.client.measurementconsumer.signMeasurementSpec
-import org.wfanet.measurement.consent.client.measurementconsumer.signRequisitionSpec
-import org.wfanet.measurement.consent.client.measurementconsumer.verifyResult
 import org.wfanet.measurement.internal.reporting.CreateReportRequest as InternalCreateReportRequest
 import org.wfanet.measurement.internal.reporting.CreateReportRequest.MeasurementKey as InternalMeasurementKey
 import org.wfanet.measurement.internal.reporting.CreateReportRequestKt.measurementKey as internalMeasurementKey
 import org.wfanet.measurement.internal.reporting.Measurement as InternalMeasurement
-import org.wfanet.measurement.internal.reporting.Measurement.Result as InternalMeasurementResult
-import org.wfanet.measurement.internal.reporting.MeasurementKt.ResultKt.frequency as internalFrequency
-import org.wfanet.measurement.internal.reporting.MeasurementKt.ResultKt.impression as internalImpression
-import org.wfanet.measurement.internal.reporting.MeasurementKt.ResultKt.reach as internalReach
-import org.wfanet.measurement.internal.reporting.MeasurementKt.ResultKt.watchDuration as internalWatchDuration
-import org.wfanet.measurement.internal.reporting.MeasurementKt.failure as internalFailure
-import org.wfanet.measurement.internal.reporting.MeasurementKt.result as internalMeasurementResult
 import org.wfanet.measurement.internal.reporting.MeasurementsGrpcKt.MeasurementsCoroutineStub as InternalMeasurementsCoroutineStub
 import org.wfanet.measurement.internal.reporting.Metric as InternalMetric
 import org.wfanet.measurement.internal.reporting.Metric.Details as InternalMetricDetails
@@ -123,10 +77,8 @@ import org.wfanet.measurement.internal.reporting.MetricKt.watchDurationParams as
 import org.wfanet.measurement.internal.reporting.PeriodicTimeInterval as InternalPeriodicTimeInterval
 import org.wfanet.measurement.internal.reporting.Report as InternalReport
 import org.wfanet.measurement.internal.reporting.ReportKt.details as internalReportDetails
-import org.wfanet.measurement.internal.reporting.ReportingSet as InternalReportingSet
 import org.wfanet.measurement.internal.reporting.ReportingSetsGrpcKt.ReportingSetsCoroutineStub as InternalReportingSetsCoroutineStub
 import org.wfanet.measurement.internal.reporting.ReportsGrpcKt.ReportsCoroutineStub as InternalReportsCoroutineStub
-import org.wfanet.measurement.internal.reporting.SetMeasurementResultRequest as SetInternalMeasurementResultRequest
 import org.wfanet.measurement.internal.reporting.StreamReportsRequest as StreamInternalReportsRequest
 import org.wfanet.measurement.internal.reporting.StreamReportsRequestKt.filter
 import org.wfanet.measurement.internal.reporting.TimeInterval as InternalTimeInterval
@@ -135,12 +87,9 @@ import org.wfanet.measurement.internal.reporting.createReportRequest as internal
 import org.wfanet.measurement.internal.reporting.getReportByIdempotencyKeyRequest
 import org.wfanet.measurement.internal.reporting.getReportRequest as getInternalReportRequest
 import org.wfanet.measurement.internal.reporting.getReportingSetRequest
-import org.wfanet.measurement.internal.reporting.measurement as internalMeasurement
 import org.wfanet.measurement.internal.reporting.metric as internalMetric
 import org.wfanet.measurement.internal.reporting.periodicTimeInterval as internalPeriodicTimeInterval
 import org.wfanet.measurement.internal.reporting.report as internalReport
-import org.wfanet.measurement.internal.reporting.setMeasurementFailureRequest as setInternalMeasurementFailureRequest
-import org.wfanet.measurement.internal.reporting.setMeasurementResultRequest as setInternalMeasurementResultRequest
 import org.wfanet.measurement.internal.reporting.streamReportsRequest as streamInternalReportsRequest
 import org.wfanet.measurement.internal.reporting.timeInterval as internalTimeInterval
 import org.wfanet.measurement.internal.reporting.timeIntervals as internalTimeIntervals
@@ -187,84 +136,6 @@ import org.wfanet.measurement.reporting.v1alpha.timeIntervals
 private const val MIN_PAGE_SIZE = 1
 private const val DEFAULT_PAGE_SIZE = 50
 private const val MAX_PAGE_SIZE = 1000
-
-private const val NUMBER_VID_BUCKETS = 300
-private const val REACH_ONLY_VID_SAMPLING_WIDTH = 3.0f / NUMBER_VID_BUCKETS
-private const val NUMBER_REACH_ONLY_BUCKETS = 16
-private val REACH_ONLY_VID_SAMPLING_START_LIST =
-  (0 until NUMBER_REACH_ONLY_BUCKETS).map { it * REACH_ONLY_VID_SAMPLING_WIDTH }
-private const val REACH_ONLY_REACH_EPSILON = 0.0041
-private const val REACH_ONLY_FREQUENCY_EPSILON = 0.0001
-private const val REACH_ONLY_MAXIMUM_FREQUENCY_PER_USER = 1
-
-private const val REACH_FREQUENCY_VID_SAMPLING_WIDTH = 5.0f / NUMBER_VID_BUCKETS
-private const val NUMBER_REACH_FREQUENCY_BUCKETS = 19
-private val REACH_FREQUENCY_VID_SAMPLING_START_LIST =
-  (0 until NUMBER_REACH_FREQUENCY_BUCKETS).map {
-    REACH_ONLY_VID_SAMPLING_START_LIST.last() +
-      REACH_ONLY_VID_SAMPLING_WIDTH +
-      it * REACH_FREQUENCY_VID_SAMPLING_WIDTH
-  }
-private const val REACH_FREQUENCY_REACH_EPSILON = 0.0033
-private const val REACH_FREQUENCY_FREQUENCY_EPSILON = 0.115
-
-private const val IMPRESSION_VID_SAMPLING_WIDTH = 62.0f / NUMBER_VID_BUCKETS
-private const val NUMBER_IMPRESSION_BUCKETS = 1
-private val IMPRESSION_VID_SAMPLING_START_LIST =
-  (0 until NUMBER_IMPRESSION_BUCKETS).map {
-    REACH_FREQUENCY_VID_SAMPLING_START_LIST.last() +
-      REACH_FREQUENCY_VID_SAMPLING_WIDTH +
-      it * IMPRESSION_VID_SAMPLING_WIDTH
-  }
-private const val IMPRESSION_EPSILON = 0.0011
-
-private const val WATCH_DURATION_VID_SAMPLING_WIDTH = 95.0f / NUMBER_VID_BUCKETS
-private const val NUMBER_WATCH_DURATION_BUCKETS = 1
-private val WATCH_DURATION_VID_SAMPLING_START_LIST =
-  (0 until NUMBER_WATCH_DURATION_BUCKETS).map {
-    IMPRESSION_VID_SAMPLING_START_LIST.last() +
-      IMPRESSION_VID_SAMPLING_WIDTH +
-      it * WATCH_DURATION_VID_SAMPLING_WIDTH
-  }
-private const val WATCH_DURATION_EPSILON = 0.001
-
-private const val DIFFERENTIAL_PRIVACY_DELTA = 1e-12
-
-private val REACH_ONLY_MEASUREMENT_SPEC = measurementSpecReachAndFrequency {
-  reachPrivacyParams = differentialPrivacyParams {
-    epsilon = REACH_ONLY_REACH_EPSILON
-    delta = DIFFERENTIAL_PRIVACY_DELTA
-  }
-  frequencyPrivacyParams = differentialPrivacyParams {
-    epsilon = REACH_ONLY_FREQUENCY_EPSILON
-    delta = DIFFERENTIAL_PRIVACY_DELTA
-  }
-  maximumFrequencyPerUser = REACH_ONLY_MAXIMUM_FREQUENCY_PER_USER
-}
-
-internal data class ReportInfo(
-  val measurementConsumerReferenceId: String,
-  val reportIdempotencyKey: String,
-  val eventGroupFilters: Map<String, String>,
-)
-
-internal data class SigningConfig(
-  val signingCertificateName: String,
-  val signingCertificateDer: ByteString,
-  val signingPrivateKey: PrivateKey,
-)
-
-internal data class WeightedMeasurementInfo(
-  val reportingMeasurementId: String,
-  val weightedMeasurement: WeightedMeasurement,
-  val timeInterval: TimeInterval,
-  var kingdomMeasurementId: String? = null,
-)
-
-internal data class SetOperationResult(
-  val weightedMeasurementInfoList: List<WeightedMeasurementInfo>,
-  val internalMetricDetails: InternalMetricDetails,
-)
 
 class ReportsService(
   private val internalReportsStub: InternalReportsCoroutineStub,
@@ -315,12 +186,13 @@ class ReportsService(
 
     if (existingInternalReport != null) return existingInternalReport.toReport()
 
-    val reportInfo: ReportInfo = buildReportInfo(request, resourceKey.measurementConsumerId)
+    val measurementInfo: MeasurementInfo =
+      buildReportInfo(request, resourceKey.measurementConsumerId)
 
     val namedSetOperationResults: Map<String, SetOperationResult> =
       compileAllSetOperations(
         request,
-        reportInfo,
+        measurementInfo,
       )
 
     val measurementConsumer =
@@ -357,19 +229,28 @@ class ReportsService(
         )
       )
 
-    createMeasurements(
-      request,
+    val measurementAgent =
+      MeasurementAgent(
+        internalReportingSetsStub,
+        internalMeasurementsStub,
+        dataProvidersStub,
+        measurementsStub,
+        certificateStub,
+        measurementConsumer,
+        apiAuthenticationKey,
+        secureRandom,
+      )
+
+    measurementAgent.createMeasurements(
       namedSetOperationResults,
-      reportInfo,
-      measurementConsumer,
-      apiAuthenticationKey,
+      measurementInfo,
       signingConfig,
     )
 
     val internalCreateReportRequest: InternalCreateReportRequest =
       buildInternalCreateReportRequest(
         request,
-        reportInfo,
+        measurementInfo,
         namedSetOperationResults,
       )
     try {
@@ -399,11 +280,11 @@ class ReportsService(
     }
   }
 
-  /** Builds a [ReportInfo] from a [CreateReportRequest]. */
+  /** Builds a [MeasurementInfo] from a [CreateReportRequest]. */
   private fun buildReportInfo(
     request: CreateReportRequest,
     measurementConsumerReferenceId: String
-  ): ReportInfo {
+  ): MeasurementInfo {
     grpcRequire(request.report.hasEventGroupUniverse()) { "EventGroupUniverse is not specified." }
     grpcRequire(request.report.metricsList.isNotEmpty()) { "Metrics in Report cannot be empty." }
     checkSetOperationNamesUniqueness(request.report.metricsList)
@@ -411,113 +292,17 @@ class ReportsService(
     val eventGroupFilters =
       request.report.eventGroupUniverse.eventGroupEntriesList.associate { it.key to it.value }
 
-    return ReportInfo(
+    return MeasurementInfo(
       measurementConsumerReferenceId,
       request.report.reportIdempotencyKey,
       eventGroupFilters,
     )
   }
 
-  /** Creates CMM public [Measurement]s and [InternalMeasurement]s from [SetOperationResult]s. */
-  private suspend fun createMeasurements(
-    request: CreateReportRequest,
-    namedSetOperationResults: Map<String, SetOperationResult>,
-    reportInfo: ReportInfo,
-    measurementConsumer: MeasurementConsumer,
-    apiAuthenticationKey: String,
-    signingConfig: SigningConfig,
-  ) = coroutineScope {
-    for (metric in request.report.metricsList) {
-      val internalMetricDetails = buildInternalMetricDetails(metric)
-
-      for (namedSetOperation in metric.setOperationsList) {
-        val setOperationId =
-          buildSetOperationId(
-            reportInfo.reportIdempotencyKey,
-            internalMetricDetails,
-            namedSetOperation.uniqueName,
-          )
-
-        val setOperationResult: SetOperationResult =
-          namedSetOperationResults[setOperationId] ?: continue
-
-        setOperationResult.weightedMeasurementInfoList.forEach { weightedMeasurementInfo ->
-          launch {
-            createMeasurement(
-              weightedMeasurementInfo,
-              reportInfo,
-              setOperationResult.internalMetricDetails,
-              measurementConsumer,
-              apiAuthenticationKey,
-              signingConfig,
-            )
-          }
-        }
-      }
-    }
-  }
-
-  /** Creates a measurement for a [WeightedMeasurement]. */
-  private suspend fun createMeasurement(
-    weightedMeasurementInfo: WeightedMeasurementInfo,
-    reportInfo: ReportInfo,
-    internalMetricDetails: InternalMetricDetails,
-    measurementConsumer: MeasurementConsumer,
-    apiAuthenticationKey: String,
-    signingConfig: SigningConfig,
-  ) {
-    val dataProviderNameToInternalEventGroupEntriesList =
-      aggregateInternalEventGroupEntryByDataProviderName(
-        weightedMeasurementInfo.weightedMeasurement.reportingSets,
-        weightedMeasurementInfo.timeInterval.toMeasurementTimeInterval(),
-        reportInfo.eventGroupFilters
-      )
-
-    val createMeasurementRequest: CreateMeasurementRequest =
-      buildCreateMeasurementRequest(
-        measurementConsumer,
-        dataProviderNameToInternalEventGroupEntriesList,
-        internalMetricDetails,
-        weightedMeasurementInfo.reportingMeasurementId,
-        apiAuthenticationKey,
-        signingConfig,
-      )
-
-    try {
-      val measurement =
-        measurementsStub
-          .withAuthenticationKey(apiAuthenticationKey)
-          .createMeasurement(createMeasurementRequest)
-      weightedMeasurementInfo.kingdomMeasurementId =
-        checkNotNull(MeasurementKey.fromName(measurement.name)).measurementId
-    } catch (e: StatusException) {
-      throw Exception(
-        "Unable to create the measurement [${createMeasurementRequest.measurement.name}].",
-        e
-      )
-    }
-
-    try {
-      internalMeasurementsStub.createMeasurement(
-        internalMeasurement {
-          this.measurementConsumerReferenceId = reportInfo.measurementConsumerReferenceId
-          this.measurementReferenceId = weightedMeasurementInfo.kingdomMeasurementId!!
-          state = InternalMeasurement.State.PENDING
-        }
-      )
-    } catch (e: StatusException) {
-      throw Exception(
-        "Unable to create the measurement [${createMeasurementRequest.measurement.name}] " +
-          "in the reporting database.",
-        e
-      )
-    }
-  }
-
   /** Compiles all [SetOperation]s and outputs each result with measurement reference ID. */
   private suspend fun compileAllSetOperations(
     request: CreateReportRequest,
-    reportInfo: ReportInfo,
+    measurementInfo: MeasurementInfo,
   ): Map<String, SetOperationResult> {
     val namedSetOperationResults = mutableMapOf<String, SetOperationResult>()
 
@@ -535,12 +320,12 @@ class ReportsService(
 
         for (namedSetOperation in metric.setOperationsList) {
           launch {
-            checkSetOperationReportingSetCoverage(namedSetOperation.setOperation, reportInfo)
+            checkSetOperationReportingSetCoverage(namedSetOperation.setOperation, measurementInfo)
           }
 
           val setOperationId =
             buildSetOperationId(
-              reportInfo.reportIdempotencyKey,
+              measurementInfo.idempotencyKey,
               internalMetricDetails,
               namedSetOperation.uniqueName
             )
@@ -587,19 +372,22 @@ class ReportsService(
   /** Checks if all reporting sets under a [SetOperation] are covered by the event filters. */
   private suspend fun checkSetOperationReportingSetCoverage(
     setOperation: SetOperation,
-    reportInfo: ReportInfo
+    measurementInfo: MeasurementInfo
   ) {
-    checkOperandReportingSetCoverage(setOperation.lhs, reportInfo)
-    checkOperandReportingSetCoverage(setOperation.rhs, reportInfo)
+    checkOperandReportingSetCoverage(setOperation.lhs, measurementInfo)
+    checkOperandReportingSetCoverage(setOperation.rhs, measurementInfo)
   }
 
   /** Checks if all reporting sets under a [Operand] are covered by the event filters. */
-  private suspend fun checkOperandReportingSetCoverage(operand: Operand, reportInfo: ReportInfo) {
+  private suspend fun checkOperandReportingSetCoverage(
+    operand: Operand,
+    measurementInfo: MeasurementInfo
+  ) {
     @Suppress("WHEN_ENUM_CAN_BE_NULL_IN_JAVA") // Proto enum fields are never null.
     when (operand.operandCase) {
       Operand.OperandCase.OPERATION ->
-        checkSetOperationReportingSetCoverage(operand.operation, reportInfo)
-      Operand.OperandCase.REPORTING_SET -> checkReportingSet(operand.reportingSet, reportInfo)
+        checkSetOperationReportingSetCoverage(operand.operation, measurementInfo)
+      Operand.OperandCase.REPORTING_SET -> checkReportingSet(operand.reportingSet, measurementInfo)
       Operand.OperandCase.OPERAND_NOT_SET -> {}
     }
   }
@@ -760,12 +548,40 @@ class ReportsService(
       )
     }
 
+    val measurementConsumer =
+      try {
+        measurementConsumersStub
+          .withAuthenticationKey(apiAuthenticationKey)
+          .getMeasurementConsumer(
+            getMeasurementConsumerRequest {
+              name = MeasurementConsumerKey(internalReport.measurementConsumerReferenceId).toName()
+            }
+          )
+      } catch (e: StatusException) {
+        throw Exception(
+          "Unable to retrieve the measurement consumer " +
+            "[${MeasurementConsumerKey(internalReport.measurementConsumerReferenceId).toName()}].",
+          e
+        )
+      }
+
+    val measurementAgent =
+      MeasurementAgent(
+        internalReportingSetsStub,
+        internalMeasurementsStub,
+        dataProvidersStub,
+        measurementsStub,
+        certificateStub,
+        measurementConsumer,
+        apiAuthenticationKey,
+        secureRandom,
+      )
+
     // Syncs measurements
-    syncMeasurements(
+    measurementAgent.syncMeasurements(
       internalReport.measurementsMap,
-      internalReport.measurementConsumerReferenceId,
-      apiAuthenticationKey,
-      principalName,
+      encryptionKeyPairStore,
+      principalName
     )
 
     return try {
@@ -786,159 +602,14 @@ class ReportsService(
     }
   }
 
-  /** Syncs [InternalMeasurement]s. */
-  private suspend fun syncMeasurements(
-    measurementsMap: Map<String, InternalMeasurement>,
-    measurementConsumerReferenceId: String,
-    apiAuthenticationKey: String,
-    principalName: String,
-  ) = coroutineScope {
-    for ((measurementReferenceId, internalMeasurement) in measurementsMap) {
-      // Measurement with SUCCEEDED state is already synced
-      if (internalMeasurement.state == InternalMeasurement.State.SUCCEEDED) continue
-
-      launch {
-        syncMeasurement(
-          measurementReferenceId,
-          measurementConsumerReferenceId,
-          apiAuthenticationKey,
-          principalName,
-        )
-      }
-    }
-  }
-
-  /** Syncs [InternalMeasurement] with the CMM [Measurement] given the measurement reference ID. */
-  private suspend fun syncMeasurement(
-    measurementReferenceId: String,
-    measurementConsumerReferenceId: String,
-    apiAuthenticationKey: String,
-    principalName: String,
-  ) {
-    val measurementResourceName =
-      MeasurementKey(measurementConsumerReferenceId, measurementReferenceId).toName()
-    val measurement =
-      try {
-        measurementsStub
-          .withAuthenticationKey(apiAuthenticationKey)
-          .getMeasurement(getMeasurementRequest { name = measurementResourceName })
-      } catch (e: StatusException) {
-        throw Exception("Unable to retrieve the measurement [$measurementResourceName].", e)
-      }
-
-    @Suppress("WHEN_ENUM_CAN_BE_NULL_IN_JAVA") // Proto enum fields are never null.
-    when (measurement.state) {
-      Measurement.State.SUCCEEDED -> {
-        // Converts a Measurement to an InternalMeasurement and store it into the database with
-        // SUCCEEDED state
-        val measurementSpec = MeasurementSpec.parseFrom(measurement.measurementSpec.data)
-        val encryptionPrivateKeyHandle =
-          encryptionKeyPairStore.getPrivateKeyHandle(
-            principalName,
-            EncryptionPublicKey.parseFrom(measurementSpec.measurementPublicKey).data
-          )
-            ?: failGrpc(Status.PERMISSION_DENIED) { "Encryption private key not found" }
-
-        val setInternalMeasurementResultRequest =
-          buildSetInternalMeasurementResultRequest(
-            measurementConsumerReferenceId,
-            measurementReferenceId,
-            measurement.resultsList,
-            encryptionPrivateKeyHandle,
-            apiAuthenticationKey,
-          )
-
-        try {
-          internalMeasurementsStub.setMeasurementResult(setInternalMeasurementResultRequest)
-        } catch (e: StatusException) {
-          throw Exception(
-            "Unable to update the measurement [$measurementResourceName] in the reporting " +
-              "database.",
-            e
-          )
-        }
-      }
-      Measurement.State.AWAITING_REQUISITION_FULFILLMENT,
-      Measurement.State.COMPUTING -> {} // No action needed
-      Measurement.State.FAILED,
-      Measurement.State.CANCELLED -> {
-        val setInternalMeasurementFailureRequest = setInternalMeasurementFailureRequest {
-          this.measurementConsumerReferenceId = measurementConsumerReferenceId
-          this.measurementReferenceId = measurementReferenceId
-          failure = measurement.failure.toInternal()
-        }
-
-        try {
-          internalMeasurementsStub.setMeasurementFailure(setInternalMeasurementFailureRequest)
-        } catch (e: StatusException) {
-          throw Exception(
-            "Unable to update the measurement [$measurementResourceName] in the reporting " +
-              "database.",
-            e
-          )
-        }
-      }
-      Measurement.State.STATE_UNSPECIFIED -> error("The measurement state should've been set.")
-      Measurement.State.UNRECOGNIZED -> error("Unrecognized measurement state.")
-    }
-  }
-
-  /** Builds a [SetInternalMeasurementResultRequest]. */
-  private suspend fun buildSetInternalMeasurementResultRequest(
-    measurementConsumerReferenceId: String,
-    measurementReferenceId: String,
-    resultsList: List<Measurement.ResultPair>,
-    privateKeyHandle: PrivateKeyHandle,
-    apiAuthenticationKey: String,
-  ): SetInternalMeasurementResultRequest {
-
-    return setInternalMeasurementResultRequest {
-      this.measurementConsumerReferenceId = measurementConsumerReferenceId
-      this.measurementReferenceId = measurementReferenceId
-      result =
-        aggregateResults(
-          resultsList
-            .map { decryptMeasurementResultPair(it, privateKeyHandle, apiAuthenticationKey) }
-            .map(Measurement.Result::toInternal)
-        )
-    }
-  }
-
-  /** Decrypts a [Measurement.ResultPair] to [Measurement.Result] */
-  private suspend fun decryptMeasurementResultPair(
-    measurementResultPair: Measurement.ResultPair,
-    encryptionPrivateKeyHandle: PrivateKeyHandle,
-    apiAuthenticationKey: String
-  ): Measurement.Result {
-    // TODO: Cache the certificate
-    val certificate =
-      try {
-        certificateStub
-          .withAuthenticationKey(apiAuthenticationKey)
-          .getCertificate(getCertificateRequest { name = measurementResultPair.certificate })
-      } catch (e: StatusException) {
-        throw Exception(
-          "Unable to retrieve the certificate [${measurementResultPair.certificate}].",
-          e
-        )
-      }
-
-    val signedResult =
-      decryptResult(measurementResultPair.encryptedResult, encryptionPrivateKeyHandle)
-    if (!verifyResult(signedResult, readCertificate(certificate.x509Der))) {
-      error("Signature of the result is invalid.")
-    }
-    return Measurement.Result.parseFrom(signedResult.data)
-  }
-
   /** Builds an [InternalCreateReportRequest] from a public [CreateReportRequest]. */
   private suspend fun buildInternalCreateReportRequest(
     request: CreateReportRequest,
-    reportInfo: ReportInfo,
+    measurementInfo: MeasurementInfo,
     namedSetOperationResults: Map<String, SetOperationResult>,
   ): InternalCreateReportRequest {
     val internalReport: InternalReport = internalReport {
-      this.measurementConsumerReferenceId = reportInfo.measurementConsumerReferenceId
+      this.measurementConsumerReferenceId = measurementInfo.measurementConsumerReferenceId
 
       @Suppress("WHEN_ENUM_CAN_BE_NULL_IN_JAVA") // Proto enum fields are never null.
       when (request.report.timeCase) {
@@ -956,23 +627,26 @@ class ReportsService(
         for (metric in request.report.metricsList) {
           launch {
             this@internalReport.metrics +=
-              buildInternalMetric(metric, reportInfo, namedSetOperationResults)
+              buildInternalMetric(metric, measurementInfo, namedSetOperationResults)
           }
         }
       }
 
       details = internalReportDetails {
-        this.eventGroupFilters.putAll(reportInfo.eventGroupFilters)
+        this.eventGroupFilters.putAll(measurementInfo.eventGroupFilters)
       }
 
-      this.reportIdempotencyKey = reportInfo.reportIdempotencyKey
+      this.reportIdempotencyKey = measurementInfo.idempotencyKey
     }
 
     return internalCreateReportRequest {
       report = internalReport
       measurements +=
         internalReport.metricsList.flatMap { internalMetric ->
-          buildInternalMeasurementKeys(internalMetric, reportInfo.measurementConsumerReferenceId)
+          buildInternalMeasurementKeys(
+            internalMetric,
+            measurementInfo.measurementConsumerReferenceId
+          )
         }
     }
   }
@@ -980,7 +654,7 @@ class ReportsService(
   /** Builds an [InternalMetric] from a public [Metric]. */
   private suspend fun buildInternalMetric(
     metric: Metric,
-    reportInfo: ReportInfo,
+    measurementInfo: MeasurementInfo,
     namedSetOperationResults: Map<String, SetOperationResult>,
   ): InternalMetric {
     return internalMetric {
@@ -990,7 +664,7 @@ class ReportsService(
         metric.setOperationsList.map { setOperation ->
           val setOperationId =
             buildSetOperationId(
-              reportInfo.reportIdempotencyKey,
+              measurementInfo.idempotencyKey,
               details,
               setOperation.uniqueName,
             )
@@ -1000,7 +674,7 @@ class ReportsService(
               val internalNamedSetOperation =
                 buildInternalNamedSetOperation(
                   setOperation,
-                  reportInfo,
+                  measurementInfo,
                   setOperationResult,
                 )
               namedSetOperations += internalNamedSetOperation
@@ -1014,7 +688,7 @@ class ReportsService(
   /** Builds an [InternalNamedSetOperation] from a public [NamedSetOperation]. */
   private suspend fun buildInternalNamedSetOperation(
     namedSetOperation: NamedSetOperation,
-    reportInfo: ReportInfo,
+    measurementInfo: MeasurementInfo,
     setOperationResult: SetOperationResult,
   ): InternalNamedSetOperation {
     return internalNamedSetOperation {
@@ -1022,7 +696,7 @@ class ReportsService(
       setOperation =
         buildInternalSetOperation(
           namedSetOperation.setOperation,
-          reportInfo.measurementConsumerReferenceId
+          measurementInfo.measurementConsumerReferenceId
         )
 
       this.measurementCalculations +=
@@ -1092,177 +766,12 @@ class ReportsService(
     }
   }
 
-  /** Builds a [CreateMeasurementRequest]. */
-  private suspend fun buildCreateMeasurementRequest(
-    measurementConsumer: MeasurementConsumer,
-    dataProviderNameToInternalEventGroupEntriesList: Map<String, List<EventGroupEntry>>,
-    internalMetricDetails: InternalMetricDetails,
-    measurementReferenceId: String,
-    apiAuthenticationKey: String,
-    signingConfig: SigningConfig,
-  ): CreateMeasurementRequest {
-    val measurementConsumerReferenceId =
-      grpcRequireNotNull(MeasurementConsumerKey.fromName(measurementConsumer.name)) {
-          "Invalid measurement consumer name [${measurementConsumer.name}]"
-        }
-        .measurementConsumerId
-
-    val measurementConsumerCertificate = readCertificate(signingConfig.signingCertificateDer)
-    val measurementConsumerSigningKey =
-      SigningKeyHandle(measurementConsumerCertificate, signingConfig.signingPrivateKey)
-    val measurementEncryptionPublicKey = measurementConsumer.publicKey.data
-
-    val measurementResourceName =
-      MeasurementKey(measurementConsumerReferenceId, measurementReferenceId).toName()
-
-    val measurement = measurement {
-      name = measurementResourceName
-      this.measurementConsumerCertificate = signingConfig.signingCertificateName
-
-      dataProviders +=
-        buildDataProviderEntries(
-          dataProviderNameToInternalEventGroupEntriesList,
-          measurementEncryptionPublicKey,
-          measurementConsumerSigningKey,
-          apiAuthenticationKey,
-        )
-
-      val unsignedMeasurementSpec: MeasurementSpec =
-        buildUnsignedMeasurementSpec(
-          measurementEncryptionPublicKey,
-          dataProviders.map { it.value.nonceHash },
-          internalMetricDetails,
-        )
-
-      this.measurementSpec =
-        signMeasurementSpec(unsignedMeasurementSpec, measurementConsumerSigningKey)
-
-      this.measurementReferenceId = measurementReferenceId
-    }
-
-    return createMeasurementRequest { this.measurement = measurement }
-  }
-
-  /** Builds a map of data provider resource name to a list of [EventGroupEntry]s. */
-  private suspend fun aggregateInternalEventGroupEntryByDataProviderName(
-    reportingSetNames: List<String>,
-    timeInterval: MeasurementTimeInterval,
-    eventGroupFilters: Map<String, String>,
-  ): Map<String, List<EventGroupEntry>> {
-    val internalReportingSetsList = mutableListOf<InternalReportingSet>()
-
-    coroutineScope {
-      for (reportingSetName in reportingSetNames) {
-        val reportingSetKey =
-          grpcRequireNotNull(ReportingSetKey.fromName(reportingSetName)) {
-            "Invalid reporting set name $reportingSetName."
-          }
-
-        launch {
-          internalReportingSetsList +=
-            try {
-              internalReportingSetsStub.getReportingSet(
-                getReportingSetRequest {
-                  measurementConsumerReferenceId = reportingSetKey.measurementConsumerId
-                  externalReportingSetId = apiIdToExternalId(reportingSetKey.reportingSetId)
-                }
-              )
-            } catch (e: StatusException) {
-              throw Exception(
-                "Unable to retrieve the reporting set [$reportingSetName] from the reporting " +
-                  "database.",
-                e
-              )
-            }
-        }
-      }
-    }
-
-    val dataProviderNameToInternalEventGroupEntriesList =
-      mutableMapOf<String, MutableList<EventGroupEntry>>()
-
-    for (internalReportingSet in internalReportingSetsList) {
-      for (eventGroupKey in internalReportingSet.eventGroupKeysList) {
-        val dataProviderName = DataProviderKey(eventGroupKey.dataProviderReferenceId).toName()
-        val eventGroupName =
-          EventGroupKey(
-              eventGroupKey.measurementConsumerReferenceId,
-              eventGroupKey.dataProviderReferenceId,
-              eventGroupKey.eventGroupReferenceId
-            )
-            .toName()
-
-        dataProviderNameToInternalEventGroupEntriesList.getOrPut(
-          dataProviderName,
-          ::mutableListOf
-        ) +=
-          eventGroupEntry {
-            key = eventGroupName
-            value = eventGroupEntryValue {
-              collectionInterval = timeInterval
-
-              val filter =
-                combineEventGroupFilters(
-                  internalReportingSet.filter,
-                  eventGroupFilters[eventGroupName]
-                )
-              if (filter != null) {
-                this.filter = requisitionSpecEventFilter { expression = filter }
-              }
-            }
-          }
-      }
-    }
-
-    return dataProviderNameToInternalEventGroupEntriesList.mapValues { it.value.toList() }.toMap()
-  }
-
-  /** Builds a list of [DataProviderEntry]s from lists of [EventGroupEntry]s. */
-  private suspend fun buildDataProviderEntries(
-    dataProviderNameToInternalEventGroupEntriesList: Map<String, List<EventGroupEntry>>,
-    measurementEncryptionPublicKey: ByteString,
-    measurementConsumerSigningKey: SigningKeyHandle,
-    apiAuthenticationKey: String,
-  ): List<DataProviderEntry> {
-    return dataProviderNameToInternalEventGroupEntriesList.map {
-      (dataProviderName, eventGroupEntriesList) ->
-      dataProviderEntry {
-        val requisitionSpec = requisitionSpec {
-          eventGroups += eventGroupEntriesList
-          this.measurementPublicKey = measurementEncryptionPublicKey
-          nonce = secureRandom.nextLong()
-        }
-
-        val dataProvider =
-          try {
-            dataProvidersStub
-              .withAuthenticationKey(apiAuthenticationKey)
-              .getDataProvider(getDataProviderRequest { name = dataProviderName })
-          } catch (e: StatusException) {
-            throw Exception("Unable to retrieve the data provider [$dataProviderName].", e)
-          }
-
-        key = dataProvider.name
-        value = dataProviderEntryValue {
-          dataProviderCertificate = dataProvider.certificate
-          dataProviderPublicKey = dataProvider.publicKey
-          encryptedRequisitionSpec =
-            encryptRequisitionSpec(
-              signRequisitionSpec(requisitionSpec, measurementConsumerSigningKey),
-              EncryptionPublicKey.parseFrom(dataProvider.publicKey.data)
-            )
-          nonceHash = hashSha256(requisitionSpec.nonce)
-        }
-      }
-    }
-  }
-
   /**
    * Check if the event groups in the public [ReportingSet] are covered by the event group universe.
    */
   private suspend fun checkReportingSet(
     reportingSetName: String,
-    reportInfo: ReportInfo,
+    measurementInfo: MeasurementInfo,
   ) {
     val reportingSetKey =
       grpcRequireNotNull(ReportingSetKey.fromName(reportingSetName)) {
@@ -1270,7 +779,7 @@ class ReportsService(
       }
 
     grpcRequire(
-      reportingSetKey.measurementConsumerId == reportInfo.measurementConsumerReferenceId
+      reportingSetKey.measurementConsumerId == measurementInfo.measurementConsumerReferenceId
     ) {
       "No access to the reporting set [$reportingSetName]."
     }
@@ -1279,7 +788,7 @@ class ReportsService(
       try {
         internalReportingSetsStub.getReportingSet(
           getReportingSetRequest {
-            this.measurementConsumerReferenceId = reportInfo.measurementConsumerReferenceId
+            this.measurementConsumerReferenceId = measurementInfo.measurementConsumerReferenceId
             externalReportingSetId = apiIdToExternalId(reportingSetKey.reportingSetId)
           }
         )
@@ -1299,53 +808,9 @@ class ReportsService(
           )
           .toName()
       val internalReportingSetDisplayName = internalReportingSet.displayName
-      grpcRequire(reportInfo.eventGroupFilters.containsKey(eventGroupName)) {
+      grpcRequire(measurementInfo.eventGroupFilters.containsKey(eventGroupName)) {
         "The event group [$eventGroupName] in the reporting set " +
           "[$internalReportingSetDisplayName] is not included in the event group universe."
-      }
-    }
-  }
-
-  /** Builds the unsigned [MeasurementSpec]. */
-  private fun buildUnsignedMeasurementSpec(
-    measurementEncryptionPublicKey: ByteString,
-    nonceHashes: List<ByteString>,
-    internalMetricDetails: InternalMetricDetails,
-  ): MeasurementSpec {
-    return measurementSpec {
-      measurementPublicKey = measurementEncryptionPublicKey
-      this.nonceHashes += nonceHashes
-
-      @Suppress("WHEN_ENUM_CAN_BE_NULL_IN_JAVA") // Proto enum fields are never null.
-      when (internalMetricDetails.metricTypeCase) {
-        InternalMetricTypeCase.REACH -> {
-          reachAndFrequency = REACH_ONLY_MEASUREMENT_SPEC
-          vidSamplingInterval = buildReachOnlyVidSamplingInterval(secureRandom)
-        }
-        InternalMetricTypeCase.FREQUENCY_HISTOGRAM -> {
-          reachAndFrequency =
-            buildReachAndFrequencyMeasurementSpec(
-              internalMetricDetails.frequencyHistogram.maximumFrequencyPerUser
-            )
-          vidSamplingInterval = buildReachAndFrequencyVidSamplingInterval(secureRandom)
-        }
-        InternalMetricTypeCase.IMPRESSION_COUNT -> {
-          impression =
-            buildImpressionMeasurementSpec(
-              internalMetricDetails.impressionCount.maximumFrequencyPerUser
-            )
-          vidSamplingInterval = buildImpressionVidSamplingInterval(secureRandom)
-        }
-        InternalMetricTypeCase.WATCH_DURATION -> {
-          duration =
-            buildDurationMeasurementSpec(
-              internalMetricDetails.watchDuration.maximumWatchDurationPerUser,
-              internalMetricDetails.watchDuration.maximumFrequencyPerUser
-            )
-          vidSamplingInterval = buildDurationVidSamplingInterval(secureRandom)
-        }
-        InternalMetricTypeCase.METRICTYPE_NOT_SET ->
-          error("Unset metric type should've already raised error.")
       }
     }
   }
@@ -1400,15 +865,6 @@ private fun buildInternalMeasurementKeys(
     }
 }
 
-/** Converts an [TimeInterval] to a [MeasurementTimeInterval] for measurement request. */
-internal fun TimeInterval.toMeasurementTimeInterval(): MeasurementTimeInterval {
-  val source = this
-  return measurementTimeInterval {
-    startTime = source.startTime
-    endTime = source.endTime
-  }
-}
-
 /** Builds a unique ID for a [SetOperation]. */
 private fun buildSetOperationId(
   reportIdempotencyKey: String,
@@ -1437,103 +893,6 @@ private fun buildMeasurementReferenceId(
 ): String {
   val rowHeader = buildRowHeader(timeInterval)
   return "$setOperationId-$rowHeader-measurement-$index"
-}
-
-/** Combines two event group filters. */
-private fun combineEventGroupFilters(filter1: String?, filter2: String?): String? {
-  if (filter1 == null) return filter2
-
-  return if (filter2 == null) filter1
-  else {
-    "($filter1) AND ($filter2)"
-  }
-}
-
-/** Builds a [VidSamplingInterval] for reach-only. */
-private fun buildReachOnlyVidSamplingInterval(secureRandom: SecureRandom): VidSamplingInterval {
-  return vidSamplingInterval {
-    // Random draw the start point from the list
-    val index = secureRandom.nextInt(NUMBER_REACH_ONLY_BUCKETS)
-    start = REACH_ONLY_VID_SAMPLING_START_LIST[index]
-    width = REACH_ONLY_VID_SAMPLING_WIDTH
-  }
-}
-
-/** Builds a [VidSamplingInterval] for reach-frequency. */
-private fun buildReachAndFrequencyVidSamplingInterval(
-  secureRandom: SecureRandom
-): VidSamplingInterval {
-  return vidSamplingInterval {
-    // Random draw the start point from the list
-    val index = secureRandom.nextInt(NUMBER_REACH_FREQUENCY_BUCKETS)
-    start = REACH_FREQUENCY_VID_SAMPLING_START_LIST[index]
-    width = REACH_FREQUENCY_VID_SAMPLING_WIDTH
-  }
-}
-
-/** Builds a [VidSamplingInterval] for impression count. */
-private fun buildImpressionVidSamplingInterval(secureRandom: SecureRandom): VidSamplingInterval {
-  return vidSamplingInterval {
-    // Random draw the start point from the list
-    val index = secureRandom.nextInt(NUMBER_IMPRESSION_BUCKETS)
-    start = IMPRESSION_VID_SAMPLING_START_LIST[index]
-    width = IMPRESSION_VID_SAMPLING_WIDTH
-  }
-}
-
-/** Builds a [VidSamplingInterval] for watch duration. */
-private fun buildDurationVidSamplingInterval(secureRandom: SecureRandom): VidSamplingInterval {
-  return vidSamplingInterval {
-    // Random draw the start point from the list
-    val index = secureRandom.nextInt(NUMBER_WATCH_DURATION_BUCKETS)
-    start = WATCH_DURATION_VID_SAMPLING_START_LIST[index]
-    width = WATCH_DURATION_VID_SAMPLING_WIDTH
-  }
-}
-
-/** Builds a [MeasurementSpec.ReachAndFrequency] for reach-frequency. */
-private fun buildReachAndFrequencyMeasurementSpec(
-  maximumFrequencyPerUser: Int
-): MeasurementSpec.ReachAndFrequency {
-  return measurementSpecReachAndFrequency {
-    reachPrivacyParams = differentialPrivacyParams {
-      epsilon = REACH_FREQUENCY_REACH_EPSILON
-      delta = DIFFERENTIAL_PRIVACY_DELTA
-    }
-    frequencyPrivacyParams = differentialPrivacyParams {
-      epsilon = REACH_FREQUENCY_FREQUENCY_EPSILON
-      delta = DIFFERENTIAL_PRIVACY_DELTA
-    }
-    this.maximumFrequencyPerUser = maximumFrequencyPerUser
-  }
-}
-
-/** Builds a [MeasurementSpec.ReachAndFrequency] for impression count. */
-private fun buildImpressionMeasurementSpec(
-  maximumFrequencyPerUser: Int
-): MeasurementSpec.Impression {
-  return measurementSpecImpression {
-    privacyParams = differentialPrivacyParams {
-      epsilon = IMPRESSION_EPSILON
-      delta = DIFFERENTIAL_PRIVACY_DELTA
-    }
-    this.maximumFrequencyPerUser = maximumFrequencyPerUser
-  }
-}
-
-/** Builds a [MeasurementSpec.ReachAndFrequency] for watch duration. */
-private fun buildDurationMeasurementSpec(
-  maximumWatchDurationPerUser: Int,
-  maximumFrequencyPerUser: Int
-): MeasurementSpec.Duration {
-  return measurementSpecDuration {
-    privacyParams = differentialPrivacyParams {
-      epsilon = WATCH_DURATION_EPSILON
-      delta = DIFFERENTIAL_PRIVACY_DELTA
-    }
-    this.maximumWatchDurationPerUser = maximumWatchDurationPerUser
-    this.maximumFrequencyPerUser = maximumFrequencyPerUser
-  }
 }
 
 /** Converts a public [SetOperation.Type] to an [InternalSetOperation.Type]. */
@@ -1625,113 +984,6 @@ private fun buildRowHeader(timeInterval: TimeInterval): String {
   val endTimeInstant =
     Instant.ofEpochSecond(timeInterval.endTime.seconds, timeInterval.endTime.nanos.toLong())
   return "$startTimeInstant-$endTimeInstant"
-}
-
-private operator fun Duration.plus(other: Duration): Duration {
-  return Durations.add(this, other)
-}
-
-/** Converts a CMM [Measurement.Failure] to an [InternalMeasurement.Failure]. */
-private fun Measurement.Failure.toInternal(): InternalMeasurement.Failure {
-  val source = this
-
-  return internalFailure {
-    @Suppress("WHEN_ENUM_CAN_BE_NULL_IN_JAVA") // Proto enum fields are never null.
-    reason =
-      when (source.reason) {
-        Measurement.Failure.Reason.REASON_UNSPECIFIED ->
-          InternalMeasurement.Failure.Reason.REASON_UNSPECIFIED
-        Measurement.Failure.Reason.CERTIFICATE_REVOKED ->
-          InternalMeasurement.Failure.Reason.CERTIFICATE_REVOKED
-        Measurement.Failure.Reason.REQUISITION_REFUSED ->
-          InternalMeasurement.Failure.Reason.REQUISITION_REFUSED
-        Measurement.Failure.Reason.COMPUTATION_PARTICIPANT_FAILED ->
-          InternalMeasurement.Failure.Reason.COMPUTATION_PARTICIPANT_FAILED
-        Measurement.Failure.Reason.UNRECOGNIZED -> InternalMeasurement.Failure.Reason.UNRECOGNIZED
-      }
-    message = source.message
-  }
-}
-
-/** Aggregate a list of [InternalMeasurementResult] to a [InternalMeasurementResult] */
-private fun aggregateResults(
-  internalResultsList: List<InternalMeasurementResult>
-): InternalMeasurementResult {
-  if (internalResultsList.isEmpty()) {
-    error("No measurement result.")
-  }
-
-  var reachValue = 0L
-  var impressionValue = 0L
-  val frequencyDistribution = mutableMapOf<Long, Double>()
-  var watchDurationValue = duration {
-    seconds = 0
-    nanos = 0
-  }
-
-  // Aggregation
-  for (result in internalResultsList) {
-    if (result.hasFrequency()) {
-      if (!result.hasReach()) {
-        error("Missing reach measurement in the Reach-Frequency measurement.")
-      }
-      for ((frequency, percentage) in result.frequency.relativeFrequencyDistributionMap) {
-        val previousTotalReachCount =
-          frequencyDistribution.getOrDefault(frequency, 0.0) * reachValue
-        val currentReachCount = percentage * result.reach.value
-        frequencyDistribution[frequency] =
-          (previousTotalReachCount + currentReachCount) / (reachValue + result.reach.value)
-      }
-    }
-    if (result.hasReach()) {
-      reachValue += result.reach.value
-    }
-    if (result.hasImpression()) {
-      impressionValue += result.impression.value
-    }
-    if (result.hasWatchDuration()) {
-      watchDurationValue += result.watchDuration.value
-    }
-  }
-
-  return internalMeasurementResult {
-    if (internalResultsList.first().hasReach()) {
-      this.reach = internalReach { value = reachValue }
-    }
-    if (internalResultsList.first().hasFrequency()) {
-      this.frequency = internalFrequency {
-        relativeFrequencyDistribution.putAll(frequencyDistribution)
-      }
-    }
-    if (internalResultsList.first().hasImpression()) {
-      this.impression = internalImpression { value = impressionValue }
-    }
-    if (internalResultsList.first().hasWatchDuration()) {
-      this.watchDuration = internalWatchDuration { value = watchDurationValue }
-    }
-  }
-}
-
-/** Converts a CMM [Measurement.Result] to an [InternalMeasurementResult]. */
-private fun Measurement.Result.toInternal(): InternalMeasurementResult {
-  val source = this
-
-  return internalMeasurementResult {
-    if (source.hasReach()) {
-      this.reach = internalReach { value = source.reach.value }
-    }
-    if (source.hasFrequency()) {
-      this.frequency = internalFrequency {
-        relativeFrequencyDistribution.putAll(source.frequency.relativeFrequencyDistributionMap)
-      }
-    }
-    if (source.hasImpression()) {
-      this.impression = internalImpression { value = source.impression.value }
-    }
-    if (source.hasWatchDuration()) {
-      this.watchDuration = internalWatchDuration { value = source.watchDuration.value }
-    }
-  }
 }
 
 /** Converts an internal [InternalReport] to a public [Report]. */
